+++ go/main.go
@@ -17,7 +17,7 @@ import (
 	"strings"
 )
 
-var version = "1.2.1"
+var version = "1.2.2"
 
 func main() {
 	if len(os.Args) < 2 {